fix valores de muestra
fix (funcion1): se corrigieron los tipos de variables de las variables numericas

se ajusto el tipo de variable de los datos numericos de int a double para evitar problemas

+++ funcion1.cs
@@ -1,8 +1,8 @@
 static void Main(string[] args){
     Console.write("hola mundo");
 
-    int valor1 = 4;
-    int valor2 = 8;
+    double valor1 = 4;
+    double valor2 = 8;
 
     int resultado = valor1 +valor2;
     Console.write(resultado);
